+++ Runtime/Attribute/InEditorAttribute.cs
@@ -1,38 +1,67 @@
 using UnityEngine;
 using System;
+using System.Net.Configuration;
 
 namespace InEditor
 {
     /// <summary>
     /// Attribute to mark field, property, method of a Type to create VisualElement at ease...
     /// </summary>
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property | AttributeTargets.Class, AllowMultiple = false)]
+    [AttributeUsage(
+        AttributeTargets.Field | AttributeTargets.Property |
+        AttributeTargets.Class, AllowMultiple = false)]
     public class InEditorAttribute : Attribute
     {
         /// <summary>
         /// Marks displaying order by this.
         /// </summary>
         public int DisplayOrder = 0;
+
         /// <summary>
         /// Marks custom name by this.
         /// </summary>
         public string DisplayName = string.Empty;
+
         /// <summary>
         /// Marks weather making display name looks nicer by this.
         /// </summary>
         public bool NicifyName = true;
+
         /// <summary>
         /// Marks displaying of the IMGUI to be a certain type.
         /// </summary>
-        public IMGUIDrawFieldEnum DesignatedIMGUI = IMGUIDrawFieldEnum.Auto;
+        public IMGUIDrawHintEnum Hint = IMGUIDrawHintEnum.Auto;
+
+        /// <summary>
+        /// Marks displaying of the value as read-only
+        /// </summary>
+        public bool Disabled = false;
+
+        /// <summary>
+        /// Marks displaying of the field to use a pre-define method in Editor or Monobehaviour.
+        /// </summary>
+        public string Method = string.Empty;
+
+        /// <summary>
+        /// Checks displaying with custom method or not.
+        /// This will disable auto drawing and data handling be careful with this.
+        /// </summary>
+        public bool HasCustomMethod
+        {
+            get { return !string.IsNullOrEmpty(Method); }
+        }
+
+
         /// <summary>
         /// Mark it! Marks a field or property to draw in BaseInEditor Editor class.
         /// Usually with an order is the best to sort drawing.
         /// </summary>
         /// <param name="order">display order, smaller first.</param>
-        public InEditorAttribute(int order = 0)
+        /// <param name="disabled">whether the field will shown as disabled</param>
+        public InEditorAttribute(int order = 0, bool disabled = false)
         {
             DisplayOrder = order;
+            Disabled = disabled;
         }
     }
 }